Extract osmacaddr functions into common_macaddress
