+++ api/v1beta1/common_webhook.go
@@ -85,23 +85,3 @@ func checkBackupOperationBlocksAction(namespace string, action shared.APIAction)
 
 	return err
 }
-
-// getAllMACReservations - get all MAC reservations in format MAC -> hostname
-func getAllMACReservations(macNodeStatus map[string]OpenStackMACNodeReservation) map[string]string {
-	ret := make(map[string]string)
-	for hostname, status := range macNodeStatus {
-		for _, mac := range status.Reservations {
-			ret[mac] = hostname
-		}
-	}
-	return ret
-}
-
-// IsUniqMAC - check if the MAC address is uniq or already present in the reservations
-func IsUniqMAC(macNodeStatus map[string]OpenStackMACNodeReservation, mac string) bool {
-	reservations := getAllMACReservations(macNodeStatus)
-	if _, ok := reservations[mac]; !ok {
-		return true
-	}
-	return false
-}
